pyproject.toml
[tool.poetry]
name = "fractal"
version = "0.1.0"
description = ""
authors = [
    "Jacopo Nespolo <jacopo.nespolo@exact-lab.it>",
    "Marco Franzon <marco.franzon@exact-lab.it>",
    "Tommaso Comparin <tommaso.comparin@exact-lab.it>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
anndata = "^0.8.0"

[tool.poetry.group.dev.dependencies]
asgi-lifespan = "^1.0.1"
pytest = "^7.1.2"
httpx = "^0.23.0"
devtools = "^0.8.0"
pytest-asyncio = "^0.18.3"
bumpver = "^2022.1116"
Pillow = "^9.1.1"
pytest-mock = "^3.8.1"
pre-commit = "^2.19.0"
coverage = "^6.4.2"
memory-profiler = "^0.60.0"


[tool.poetry.group.server.dependencies]
fastapi = "^0.78.0"
sqlmodel = "^0.0.6"
aiosqlite = "^0.17.0"
fastapi-users = {extras = ["oauth"], version = "^10.1.1"}
fastapi-users-db-sqlmodel = "^0.2.0"
alembic = "^1.8.0"
uvicorn = "^0.18.2"

[tool.poetry.group.runner.dependencies]
parsl = {extras = ["monitoring"], git="https://github.com/fractal-analytics-platform/parsl.git", branch="1.2.1-unofficial"}
dask = "^2022.6.1"
scikit-image = "^0.19.3"
cellpose = "2"
ome-zarr = "^0.5.1"
pandas = "^1.2.0"
imagecodecs = "^2022.2.22"
graphviz = "^0.20"
defusedxml = "^0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.bumpver]
current_version = "0.1.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'version = "{version}"',
]
"fractal/server/config.py" = [
    '__VERSION__ = "{version}"'
]
"README.md" = [
    "{version}",
    "{pep440_version}",
]

[tool.poetry.scripts]
server = "scripts.testserver:run_testserver"
client = "scripts.client:run"
